back/back/requirements.txt
Django
django-elasticsearch-dsl-drf
django-elasticsearch-dsl
djangorestframework
django-cors-headers
drf_yasg